requirements.txt
# Core AI and ML libraries (open-source stack)
transformers>=4.35.0
torch>=2.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# Document processing
pypdf2>=3.0.0
pdf2image>=1.16.3
pillow>=10.1.0
python-docx>=1.1.0

# OCR capabilities
pytesseract>=0.3.10
easyocr>=1.7.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
